Upgrade to 12.1.8

+++ ShaderLibrary/Input.hlsl
@@ -14,6 +14,7 @@
 #include "Packages/com.unity.render-pipelines.universal/ShaderLibrary/ShaderTypes.cs.hlsl"
 #include "Packages/com.unity.render-pipelines.universal/ShaderLibrary/Deprecated.hlsl"
 
+// Must match: UniversalRenderPipeline.maxVisibleAdditionalLights
 #if defined(SHADER_API_MOBILE) && (defined(SHADER_API_GLES) || defined(SHADER_API_GLES30))
     #define MAX_VISIBLE_LIGHTS 16
 #elif defined(SHADER_API_MOBILE) || (defined(SHADER_API_GLCORE) && !defined(SHADER_API_SWITCH)) || defined(SHADER_API_GLES) || defined(SHADER_API_GLES3) // Workaround because SHADER_API_GLCORE is also defined when SHADER_API_SWITCH is
@@ -35,7 +36,7 @@ struct InputData
 {
     float3  positionWS;
     float4  positionCS;
-    half3   normalWS;
+    float3   normalWS;
     half3   viewDirectionWS;
     float4  shadowCoord;
     half    fogCoord;
